+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Prevent a crash that could occur when an error is raised while another event is being processed ([#789](https://github.com/getsentry/sentry-godot/pull/789))
 
+### Other changes
+
+- Include the license file in the released addon and demo-project packages ([#790](https://github.com/getsentry/sentry-godot/pull/790))
+
 ### Dependencies
 
 - Bump Sentry JavaScript from v10.61.0 to v10.62.0 ([#786](https://github.com/getsentry/sentry-godot/pull/786))